requirements.txt
celery==3.0.5
Django>=1.4.1
django-admin-sortable==1.3.7
django-celery==3.0.4
django-gravatar==0.1.0
django-tinymce==1.5.1b2
django-userena==1.1.2
django-voting==0.2
django-tastypie==0.9.11
South==0.7.3
psycopg2==2.4.2
pymongo==2.2.1
djangosaml2==0.7.0
PIL>=1.1.7
django_compressor==1.1.2
